Adds Club Greensock code.
Greensock provides examples of their premium plugins on Code Pen,
available at https://codepen.io/GreenSock/pen/OPqpRJ . That code
has been included in this repo as well, slightly modified to
allow it to run on github pages. While I'm not sure if this
is within their Licensing Terms, this is for TESTING PURPOSES
ONLY, and is strictly for learning. This is not to be used
in any commercial products.